Fix unit tests for x86_64 where signed long long is really large.
On 64bit architecture, the signed long long in C will be a value
from -9223372036854775807 to 9223372036854775807. Which is an order
of magnitude bigger than signed long on 64bit or, signed long long
on 32bit architecture.

This fix was possible thanks to all the nice people at Ooyala who
kindly provided an x86_64 system to test on.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>